Changed constant factor to number of new messages
Before we were using a constant factor to divide the reset time adjustment. This had to be set by the user, and experiments showed that the optimal value depended on the number swarm density and the period. The number of messages received by each robot within a period is a function of these two factors. Now, each robot keeps track of how many messages it has received in a given period (num_messages), and then uses that to adjust the reset time. (2 * num_messages)  was experimentally determined to be a good fit. + 1 so that we never divide by 0.